requirements.txt
alembic==1.7.5
anyio==3.6.1
appdirs==1.4.4
asgiref==3.4.1
async-generator==1.10
attrs==21.4.0
boto3==1.24.17
botocore==1.27.17
cachetools==5.2.0
certifi==2021.10.8
charset-normalizer==2.0.9
Click==7.0
distlib==0.3.4
Django==3.2.8
filelock==3.4.2
Flask==1.1.1
Flask-Login==0.5.0
Flask-Migrate==3.1.0
Flask-PyMongo==2.3.0
Flask-SQLAlchemy==2.5.1
google-api-core==2.8.2
google-api-python-client==2.58.0
google-auth==2.11.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.4
greenlet==1.1.2
h11==0.12.0
html-testRunner==1.2.1
httpcore==0.15.0
httplib2==0.20.4
httpx==0.23.0
idna==3.3
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==2.10.1
jmespath==1.0.1
Mako==1.1.6
MarkupSafe==2.0.1
newsapi-python==0.2.6
oauthlib==3.1.0
outcome==1.1.0
packaging==21.3
pep8==1.7.1
pluggy==1.0.0
protobuf==4.21.5
psycopg2-binary==2.8.6
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==1.7.1
pymongo==4.0.1
pyparsing==3.0.9
pytest==7.1.2
python-dateutil==2.7.3
pytz==2021.3
requests==2.27.0
rfc3986==1.5.0
rsa==4.9
s3transfer==0.6.0
selenium==4.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.27
sqlparse==0.4.2
tomli==2.0.1
trio==0.19.0
trio-websocket==0.9.2
uritemplate==4.1.1
urllib3==1.26.7
virtualenv==20.0.17
Werkzeug==0.16.1
wsproto==1.0.0
zipp==3.7.0
